PPC-3: Working client<->server rpc communication
